explorerService: consolidate bulkFileService apply in one location
