+++ src/vs/workbench/contrib/files/browser/fileActions.ts
@@ -12,7 +12,7 @@ import { URI } from 'vs/base/common/uri';
 import { toErrorMessage } from 'vs/base/common/errorMessage';
 import { Action } from 'vs/base/common/actions';
 import { DisposableStore, dispose, IDisposable, toDisposable } from 'vs/base/common/lifecycle';
-import { VIEWLET_ID, IExplorerService, IFilesConfiguration, VIEW_ID } from 'vs/workbench/contrib/files/common/files';
+import { VIEWLET_ID, IFilesConfiguration, VIEW_ID } from 'vs/workbench/contrib/files/common/files';
 import { ByteSize, IFileService, IFileStatWithMetadata } from 'vs/platform/files/common/files';
 import { EditorResourceAccessor, SideBySideEditor } from 'vs/workbench/common/editor';
 import { ExplorerViewPaneContainer } from 'vs/workbench/contrib/files/browser/explorerViewlet';
@@ -52,7 +52,8 @@ import { IProgressService, IProgressStep, ProgressLocation } from 'vs/platform/p
 import { CancellationTokenSource } from 'vs/base/common/cancellation';
 import { ILogService } from 'vs/platform/log/common/log';
 import { IUriIdentityService } from 'vs/workbench/services/uriIdentity/common/uriIdentity';
-import { IBulkEditService, ResourceFileEdit } from 'vs/editor/browser/services/bulkEditService';
+import { ResourceFileEdit } from 'vs/editor/browser/services/bulkEditService';
+import { IExplorerService } from 'vs/workbench/contrib/files/browser/files';
 
 export const NEW_FILE_COMMAND_ID = 'explorer.newFile';
 export const NEW_FILE_LABEL = nls.localize('newFile', "New File");
@@ -116,7 +117,7 @@ export class NewFolderAction extends Action {
 	}
 }
 
-async function deleteFiles(explorerService: IExplorerService, bulkEditService: IBulkEditService, workingCopyFileService: IWorkingCopyFileService, dialogService: IDialogService, configurationService: IConfigurationService, elements: ExplorerItem[], useTrash: boolean, skipConfirm = false): Promise<void> {
+async function deleteFiles(explorerService: IExplorerService, workingCopyFileService: IWorkingCopyFileService, dialogService: IDialogService, configurationService: IConfigurationService, elements: ExplorerItem[], useTrash: boolean, skipConfirm = false): Promise<void> {
 	let primaryButton: string;
 	if (useTrash) {
 		primaryButton = isWindows ? nls.localize('deleteButtonLabelRecycleBin', "&&Move to Recycle Bin") : nls.localize({ key: 'deleteButtonLabelTrash', comment: ['&& denotes a mnemonic'] }, "&&Move to Trash");
@@ -220,10 +221,7 @@ async function deleteFiles(explorerService: IExplorerService, bulkEditService: I
 	// Call function
 	try {
 		const resourceFileEdits = distinctElements.map(e => new ResourceFileEdit(e.resource, undefined, { recursive: true, folder: e.isDirectory, skipTrashBin: !useTrash, maxSize: MAX_UNDO_FILE_SIZE }));
-		await bulkEditService.apply(resourceFileEdits, {
-			undoRedoSource: explorerService.undoRedoSource,
-			label: distinctElements.length > 1 ? nls.localize('deleteBulkEdit', "Delete {0} files", distinctElements.length) : nls.localize('deleteFileBulkEdit', "Delete {0}", distinctElements[0].name)
-		});
+		await explorerService.applyBulkEdit(resourceFileEdits, distinctElements.length > 1 ? nls.localize('deleteBulkEdit', "Delete {0} files", distinctElements.length) : nls.localize('deleteFileBulkEdit', "Delete {0}", distinctElements[0].name));
 	} catch (error) {
 
 		// Handle error to delete file(s) from a modal confirmation dialog
@@ -253,7 +251,7 @@ async function deleteFiles(explorerService: IExplorerService, bulkEditService: I
 
 			skipConfirm = true;
 
-			return deleteFiles(explorerService, bulkEditService, workingCopyFileService, dialogService, configurationService, elements, useTrash, skipConfirm);
+			return deleteFiles(explorerService, workingCopyFileService, dialogService, configurationService, elements, useTrash, skipConfirm);
 		}
 	}
 }
@@ -875,7 +873,6 @@ async function openExplorerAndCreate(accessor: ServicesAccessor, isFolder: boole
 	const viewsService = accessor.get(IViewsService);
 	const notificationService = accessor.get(INotificationService);
 	const commandService = accessor.get(ICommandService);
-	const bulkEditService = accessor.get(IBulkEditService);
 
 	const view = await viewsService.openView(VIEW_ID, true);
 	if (!view) {
@@ -907,10 +904,7 @@ async function openExplorerAndCreate(accessor: ServicesAccessor, isFolder: boole
 	const onSuccess = async (value: string): Promise<void> => {
 		try {
 			const resourceToCreate = resources.joinPath(folder.resource, value);
-			await bulkEditService.apply([new ResourceFileEdit(undefined, resourceToCreate, { folder: isFolder })], {
-				undoRedoSource: explorerService.undoRedoSource,
-				label: nls.localize('newBulkEdit', "New {0}", value)
-			});
+			await explorerService.applyBulkEdit([new ResourceFileEdit(undefined, resourceToCreate, { folder: isFolder })], nls.localize('newBulkEdit', "New {0}", value));
 			await refreshIfSeparator(value, explorerService);
 
 			if (isFolder) {
@@ -951,7 +945,6 @@ CommandsRegistry.registerCommand({
 
 export const renameHandler = async (accessor: ServicesAccessor) => {
 	const explorerService = accessor.get(IExplorerService);
-	const bulkEditService = accessor.get(IBulkEditService);
 	const notificationService = accessor.get(INotificationService);
 
 	const stats = explorerService.getContext(false);
@@ -968,10 +961,7 @@ export const renameHandler = async (accessor: ServicesAccessor) => {
 				const targetResource = resources.joinPath(parentResource, value);
 				if (stat.resource.toString() !== targetResource.toString()) {
 					try {
-						await bulkEditService.apply([new ResourceFileEdit(stat.resource, targetResource)], {
-							undoRedoSource: explorerService.undoRedoSource,
-							label: nls.localize('renameBulkEdit', "Rename {0} to {1}", stat.name, value)
-						});
+						await explorerService.applyBulkEdit([new ResourceFileEdit(stat.resource, targetResource)], nls.localize('renameBulkEdit', "Rename {0} to {1}", stat.name, value));
 						await refreshIfSeparator(value, explorerService);
 					} catch (e) {
 						notificationService.error(e);
@@ -987,7 +977,7 @@ export const moveFileToTrashHandler = async (accessor: ServicesAccessor) => {
 	const explorerService = accessor.get(IExplorerService);
 	const stats = explorerService.getContext(true).filter(s => !s.isRoot);
 	if (stats.length) {
-		await deleteFiles(accessor.get(IExplorerService), accessor.get(IBulkEditService), accessor.get(IWorkingCopyFileService), accessor.get(IDialogService), accessor.get(IConfigurationService), stats, true);
+		await deleteFiles(accessor.get(IExplorerService), accessor.get(IWorkingCopyFileService), accessor.get(IDialogService), accessor.get(IConfigurationService), stats, true);
 	}
 };
 
@@ -996,7 +986,7 @@ export const deleteFileHandler = async (accessor: ServicesAccessor) => {
 	const stats = explorerService.getContext(true).filter(s => !s.isRoot);
 
 	if (stats.length) {
-		await deleteFiles(accessor.get(IExplorerService), accessor.get(IBulkEditService), accessor.get(IWorkingCopyFileService), accessor.get(IDialogService), accessor.get(IConfigurationService), stats, false);
+		await deleteFiles(accessor.get(IExplorerService), accessor.get(IWorkingCopyFileService), accessor.get(IDialogService), accessor.get(IConfigurationService), stats, false);
 	}
 };
 
@@ -1024,7 +1014,6 @@ const downloadFileHandler = (accessor: ServicesAccessor) => {
 	const logService = accessor.get(ILogService);
 	const fileService = accessor.get(IFileService);
 	const fileDialogService = accessor.get(IFileDialogService);
-	const bulkEditService = accessor.get(IBulkEditService);
 	const explorerService = accessor.get(IExplorerService);
 	const progressService = accessor.get(IProgressService);
 
@@ -1243,10 +1232,7 @@ const downloadFileHandler = (accessor: ServicesAccessor) => {
 				});
 
 				if (destination) {
-					await bulkEditService.apply([new ResourceFileEdit(explorerItem.resource, destination, { overwrite: true, copy: true })], {
-						undoRedoSource: explorerService.undoRedoSource,
-						label: nls.localize('downloadBulkEdit', "Download {0}", explorerItem.name)
-					});
+					await explorerService.applyBulkEdit([new ResourceFileEdit(explorerItem.resource, destination, { overwrite: true, copy: true })], nls.localize('downloadBulkEdit', "Download {0}", explorerItem.name));
 				} else {
 					cts.cancel(); // User canceled a download. In case there were multiple files selected we should cancel the remainder of the prompts #86100
 				}
@@ -1268,7 +1254,6 @@ export const pasteFileHandler = async (accessor: ServicesAccessor) => {
 	const explorerService = accessor.get(IExplorerService);
 	const fileService = accessor.get(IFileService);
 	const notificationService = accessor.get(INotificationService);
-	const bulkEditService = accessor.get(IBulkEditService);
 	const editorService = accessor.get(IEditorService);
 	const configurationService = accessor.get(IConfigurationService);
 	const uriIdentityService = accessor.get(IUriIdentityService);
@@ -1303,16 +1288,10 @@ export const pasteFileHandler = async (accessor: ServicesAccessor) => {
 		// Move/Copy File
 		if (pasteShouldMove) {
 			const resourceFileEdits = sourceTargetPairs.map(pair => new ResourceFileEdit(pair.source, pair.target));
-			await bulkEditService.apply(resourceFileEdits, {
-				undoRedoSource: explorerService.undoRedoSource,
-				label: sourceTargetPairs.length > 1 ? nls.localize('moveBulkEdit', "Move {0} files", sourceTargetPairs.length) : nls.localize('moveFileBulkEdit', "Move {0}", resources.basenameOrAuthority(sourceTargetPairs[0].target))
-			});
+			await explorerService.applyBulkEdit(resourceFileEdits, sourceTargetPairs.length > 1 ? nls.localize('moveBulkEdit', "Move {0} files", sourceTargetPairs.length) : nls.localize('moveFileBulkEdit', "Move {0}", resources.basenameOrAuthority(sourceTargetPairs[0].target)));
 		} else {
 			const resourceFileEdits = sourceTargetPairs.map(pair => new ResourceFileEdit(pair.source, pair.target, { copy: true }));
-			await bulkEditService.apply(resourceFileEdits, {
-				undoRedoSource: explorerService.undoRedoSource,
-				label: sourceTargetPairs.length > 1 ? nls.localize('copyBulkEdit', "Copy {0} files", sourceTargetPairs.length) : nls.localize('copyFileBulkEdit', "Copy {0}", resources.basenameOrAuthority(sourceTargetPairs[0].target))
-			});
+			await explorerService.applyBulkEdit(resourceFileEdits, sourceTargetPairs.length > 1 ? nls.localize('copyBulkEdit', "Copy {0} files", sourceTargetPairs.length) : nls.localize('copyFileBulkEdit', "Copy {0}", resources.basenameOrAuthority(sourceTargetPairs[0].target)));
 		}
 
 		if (sourceTargetPairs.length >= 1) {
