+++ src/vs/workbench/contrib/files/browser/files.ts
@@ -5,7 +5,7 @@
 
 import { URI } from 'vs/base/common/uri';
 import { IListService } from 'vs/platform/list/browser/listService';
-import { OpenEditor, IExplorerService } from 'vs/workbench/contrib/files/common/files';
+import { OpenEditor, SortOrder } from 'vs/workbench/contrib/files/common/files';
 import { EditorResourceAccessor, SideBySideEditor, IEditorIdentifier } from 'vs/workbench/common/editor';
 import { List } from 'vs/base/browser/ui/list/listWidget';
 import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
@@ -13,6 +13,51 @@ import { ExplorerItem } from 'vs/workbench/contrib/files/common/explorerModel';
 import { coalesce } from 'vs/base/common/arrays';
 import { AsyncDataTree } from 'vs/base/browser/ui/tree/asyncDataTree';
 import { IEditorGroupsService } from 'vs/workbench/services/editor/common/editorGroupsService';
+import { IEditableData } from 'vs/workbench/common/views';
+import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
+import { ResourceFileEdit } from 'vs/editor/browser/services/bulkEditService';
+
+
+export interface IExplorerService {
+	readonly _serviceBrand: undefined;
+	readonly roots: ExplorerItem[];
+	readonly sortOrder: SortOrder;
+
+	getContext(respectMultiSelection: boolean): ExplorerItem[];
+	hasViewFocus(): boolean;
+	setEditable(stat: ExplorerItem, data: IEditableData | null): Promise<void>;
+	getEditable(): { stat: ExplorerItem, data: IEditableData } | undefined;
+	getEditableData(stat: ExplorerItem): IEditableData | undefined;
+	// If undefined is passed checks if any element is currently being edited.
+	isEditable(stat: ExplorerItem | undefined): boolean;
+	findClosest(resource: URI): ExplorerItem | null;
+	refresh(): Promise<void>;
+	setToCopy(stats: ExplorerItem[], cut: boolean): Promise<void>;
+	isCut(stat: ExplorerItem): boolean;
+	applyBulkEdit(edit: ResourceFileEdit[], label: string): Promise<void>;
+
+	/**
+	 * Selects and reveal the file element provided by the given resource if its found in the explorer.
+	 * Will try to resolve the path in case the explorer is not yet expanded to the file yet.
+	 */
+	select(resource: URI, reveal?: boolean | string): Promise<void>;
+
+	registerView(contextAndRefreshProvider: IExplorerView): void;
+}
+
+export const IExplorerService = createDecorator<IExplorerService>('explorerService');
+
+export interface IExplorerView {
+	getContext(respectMultiSelection: boolean): ExplorerItem[];
+	refresh(recursive: boolean, item?: ExplorerItem): Promise<void>;
+	selectResource(resource: URI | undefined, reveal?: boolean | string): Promise<void>;
+	setTreeInput(): Promise<void>;
+	itemsCopied(tats: ExplorerItem[], cut: boolean, previousCut: ExplorerItem[] | undefined): void;
+	setEditable(stat: ExplorerItem, isEditing: boolean): Promise<void>;
+	focusNeighbourIfItemFocused(item: ExplorerItem): void;
+	isItemVisible(item: ExplorerItem): boolean;
+	hasFocus(): boolean;
+}
 
 function getFocus(listService: IListService): unknown | undefined {
 	let list = listService.lastFocusedList;
